refactor(router & core): update router & core
